Moved cache controller into controllers package
